FileFilter: add line/column number to file urls; also match absolute paths
This is hardcoded to use Kate at the moment since there is no universal way
to tell a text editor about line/column on the cli, each editor seems to
have its own scheme; hopefully a profile option can be added to let the
user specify the command to use to open local text files, e.g.
/usr/bin/kate --line LINE_NUMBER --column COLUMN_NUMBER
/usr/bin/gedit +LINE_NUMBER:COLUMN_NUMBER

then we replace LINE_NUMBER and COLUMN_NUMBER to create the cmd to pass on
to KIO::CommadLauncherJob.

Also change the regex to also match absolute file paths, this is useful in
general and also compiler output usually has absolute paths.